Working version

+++ ATP2018.SpamClassification/Group.cs
@@ -5,11 +5,11 @@ namespace ATP2018.SpamClassification
 {
     public class Group
     {
-        public float Proportion { get; }
+        public double Proportion { get; }
 
         public Dictionary<string, double> TokenFrequencies { get; }
 
-        public Group(float proportion, Dictionary<string, double> tokenFrequencies)
+        public Group(double proportion, Dictionary<string, double> tokenFrequencies)
         {
             Proportion = proportion;
             TokenFrequencies = tokenFrequencies ?? throw new ArgumentNullException(nameof(tokenFrequencies));
